Fix more bugs, improve fuzzer.
