+++ Cargo.toml
@@ -13,4 +13,3 @@ license = "MIT"
 
 [dependencies]
 urlencoding = "2.1.2"
-regex = "1.7.3"
